+++ packages/gatsby-plugin-cms/src/index.ts
@@ -7,7 +7,7 @@ export interface Schema extends JSONSchema6 {
 
 export type Schemas = Record<string, Schema>
 
-interface ContentType {
+export interface ContentType {
   name: string
   extraBlocks: Record<string, Schemas>
   beforeBlocks: Schemas
