--- packages/gatsby-plugin-cms/src/native-types.ts
@@ -1,12 +1,0 @@
-export {
-  Collection,
-  isCategoryCollection,
-  isBrandCollection,
-  isClusterCollection,
-} from './node-api/catalog'
-export type {
-  ICollection,
-  ICategoryCollection,
-  IClusterCollection,
-  IBrandCollection,
-} from './node-api/catalog'
