+++ packages/gatsby-plugin-cms/src/native-types/contentTypes/plp.ts
@@ -1,0 +1,21 @@
+import { Collection } from '../blocks/collection'
+import type { ContentType } from '../../index'
+
+export const PLP = ({
+  extraBlocks = {},
+  beforeBlocks = {},
+  afterBlocks = {},
+}: Partial<Omit<ContentType, 'name'>>) => ({
+  plp: {
+    name: 'PLP',
+    extraBlocks: {
+      ...extraBlocks,
+      Parameters: {
+        ...extraBlocks.Parameters,
+        Collection,
+      },
+    },
+    beforeBlocks,
+    afterBlocks,
+  },
+})
