+++ packages/gatsby-plugin-cms/src/native-types/index.ts
@@ -1,0 +1,7 @@
+export { PLP } from './contentTypes/plp'
+
+export { Seo } from './blocks/seo'
+export type { ISeo } from './blocks/seo'
+
+export { Sort } from './blocks/sort'
+export type { ISort } from './blocks/sort'
