+++ packages/gatsby-plugin-cms/src/node-api/catalog/index.ts
@@ -1,153 +1,67 @@
-import type { Schema } from '../../index'
+import type { ParentSpanPluginArgs } from 'gatsby'
 
-type Sort =
-  | '' // 'Relevance',
-  | 'price:desc' // 'Price: High to Low',
-  | 'price:asc' // 'Price: Low to High',
-  | 'orders:desc' // 'Sales',
-  | 'name:desc' // 'Name, descending',
-  | 'name:asc' // 'Name, ascending',
-  | 'release:desc' // 'Release date',
-  | 'discount:desc' // 'Discount',
+import { nodeId } from '../cms/sourceNode'
+import type { RemotePageContent } from '../cms/types'
+import type {
+  ICollection,
+  ICategoryCollection,
+  IBrandCollection,
+  IClusterCollection,
+} from '../../native-types/blocks/collection'
+import {
+  isCategoryCollection,
+  isBrandCollection,
+  isClusterCollection,
+} from '../../native-types/blocks/collection'
 
-export interface ICategoryCollection {
-  sort: Sort
-  categoryId: string
-}
+export type WithPLP<T> = T & { plp: string }
 
-export interface IBrandCollection {
-  sort: Sort
-  brandId: string
-}
+export const getCollectionsFromPageContent = (
+  gatsbyApi: ParentSpanPluginArgs,
+  nodes: RemotePageContent[]
+) => {
+  const collectionBlocks: Array<WithPLP<ICollection>> = []
 
-export interface IClusterCollection {
-  seo: {
-    slug: string
-    title: string
-    description: string
+  for (const node of nodes) {
+    // We only allow plp content types
+    if (node.type !== 'plp') {
+      continue
+    }
+
+    for (const extraBlock of node.extraBlocks) {
+      const block = extraBlock.blocks.find((x) => x.name === 'Collection')
+
+      if (block) {
+        const props = (block.props as unknown) as ICollection
+
+        collectionBlocks.push({
+          ...props,
+          plp: gatsbyApi.createNodeId(nodeId(node)),
+        })
+      }
+    }
   }
-  sort: Sort
-  clusterId: string
+
+  return collectionBlocks
 }
 
-/**
- * Definition of a Collection in the CMS
- */
-export type ICollection =
-  | ICategoryCollection
-  | IBrandCollection
-  | IClusterCollection
-
-export const isCategoryCollection = (
-  x: ICollection
-): x is ICategoryCollection => typeof (x as any).categoryId === 'string'
-
-export const isBrandCollection = (x: ICollection): x is IBrandCollection =>
-  typeof (x as any).brandId === 'string'
-
-export const isClusterCollection = (x: ICollection): x is IClusterCollection =>
-  typeof (x as any).clusterId === 'string'
-
-const SeoSchema = {
-  type: 'object',
-  title: 'Seo',
-  widget: {
-    'ui:ObjectFieldTemplate': 'GoogleSeoPreview',
-  },
-  required: ['title', 'description', 'slug'],
-  properties: {
-    title: {
-      type: 'string',
-      title: 'Title',
-      description:
-        'Appears in the browser tab and is suggested for search engines',
-      default: 'Page title',
-    },
-    slug: {
-      type: 'string',
-      title: 'URL slug',
-      description: "Final part of the page's address. No spaces allowed.",
-      default: '/path/to/page',
-      pattern: '^/([a-zA-Z0-9]|-|/|_)*',
-    },
-    description: {
-      type: 'string',
-      title: 'Description (Meta description)',
-      description: 'Suggested for search engines',
-      default: 'Page description',
-    },
-  },
-} as Schema
-
-const SortSchema = {
-  title: 'Default ordering',
-  type: 'string',
-  default: '""',
-  enum: [
-    '""',
-    'discount:desc',
-    'release:desc',
-    'name:asc',
-    'name:desc',
-    'orders:desc',
-    'price:asc',
-    'price:desc',
-  ],
-  enumNames: [
-    'Relevance',
-    'Discount',
-    'Release date',
-    'Name, ascending',
-    'Name, descending',
-    'Sales',
-    'Price: Low to High',
-    'Price: High to Low',
-  ],
-} as Schema
-
-export const Collection = {
-  title: 'Collection',
-  description: 'Definition of a Collection for the CMS',
-  oneOf: [
-    {
-      title: 'Category',
-      description: 'Configure a Category',
-      type: 'object',
-      required: ['categoryId', 'sort'],
-      properties: {
-        categoryId: {
-          title: 'Category ID',
-          type: 'string',
-        },
-        sort: SortSchema,
-      },
-    },
-    {
-      title: 'Brand',
-      description: 'Configure a Brand',
-      type: 'object',
-      required: ['brandId', 'sort'],
-      properties: {
-        brandId: {
-          title: 'Brand ID',
-          type: 'string',
-        },
-        sort: SortSchema,
-      },
-    },
-    {
-      title: 'Collection',
-      description: 'Configure a Collection',
-      type: 'object',
-      required: ['clusterId', 'sort', 'seo'],
-      properties: {
-        clusterId: {
-          title: 'Collection ID',
-          type: 'string',
-        },
-        sort: SortSchema,
-        seo: SeoSchema,
-      },
-    },
-  ],
-} as Schema
+export const splitCollections = (collections: Array<WithPLP<ICollection>>) => ({
+  categories: collections
+    .filter((x): x is WithPLP<ICategoryCollection> => isCategoryCollection(x))
+    .reduce(
+      (acc, curr) => ({ ...acc, [curr.categoryId]: curr }),
+      {} as Record<string, WithPLP<ICategoryCollection>>
+    ),
+  brands: collections
+    .filter((x): x is WithPLP<IBrandCollection> => isBrandCollection(x))
+    .reduce(
+      (acc, curr) => ({ ...acc, [curr.brandId]: curr }),
+      {} as Record<string, WithPLP<IBrandCollection>>
+    ),
+  clusters: collections
+    .filter((x): x is WithPLP<IClusterCollection> => isClusterCollection(x))
+    .reduce(
+      (acc, curr) => ({ ...acc, [curr.clusterId]: curr }),
+      {} as Record<string, WithPLP<IClusterCollection>>
+    ),
+})
